This is a continuation of https://trac.wxwidgets.org/ticket/14403 Namely, if we render wxBitmaps with an alpha channel into a wxDataViewCtrl using a wxDataViewCustomRenderer derived class, an extremely slow bit of code kicks in, where the alpha of the whole double buffer gets cleared at each bitmap render. This makes our application crawl once we render a bitmap or two at each line of the control on a high resolution display. It happens, that wxDataViewCtrl paints into a double buffer, and the double buffer has no alpha channel assigned.
Vadim's answer:
I think the solution is to either use alpha channel for the backing store
bitmap or indeed make it 24bpp instead of 32bpp. They should both work the
same, but I'm not sure which one would be more efficient. If you already
have a possibility to do it, perhaps you could try testing this?

If you mean awi's suggestion at

        https://trac.wxwidgets.org/ticket/14403#comment:45

then it rather creates a 24bpp bitmap, without any alpha channel at all or
even any possibility to have it -- which prevents AlphaBlt() from clearing
it.

The alternative would be to create a normal 32bpp bitmap and call
UseAlpha() on it as AlphaBlt() considers that alpha values are already
correct and does nothing in this case Too.

Naively, I'd expect
UseAlpha() to be faster because blitting between bitmaps using the same
depth should be a faster path, but it would be much better to measure this.
If you could please do this and make a patch/PR changing it to use the
faster version, it would be very welcome!