Disabled assert in GetInfoFromLCID() if win32 locales do not match C locales.
It is an unfortunate situation, where we need to switch the locales
to use decimal point as a separator until we adapt all the back-end code
to switch the locales for importing  / exporting various files.
Until then, we rather switched the numeric locales to "C" everywhere.
Unfortonately there are also numeric locales defined at Win32, which
cannot be changed (they can, but system wise). We tested, that we can
safely disable this wxWidgets assert and the application will still
work correctly.